If a pig is on the edge, prepend or append column
We ensure there is a column of primary blocks to the immediate left and
to the immediate right of a pig. However, if a pig is on the edge, this
is not possible since there isn't a space to prepend or append a column.
Hence, in such cases, we prepend or append a whole new column to the
list of blocks.

However, it doesn't work properly because the platform blocks that would
normally exist between the blocks of this new column does not exist and
hence, as soon as the level starts the blocks of the new column start to
freefall.